+++ src/main/php/web/io/TestOutput.class.php
@@ -1,7 +1,26 @@
 <?php namespace web\io;
 
+use lang\XPClass;
+
 class TestOutput extends Output {
   private $bytes;
+  private $streaming;
+
+  /** Create a new Test Output */
+  public function __construct() {
+    $this->streaming= new XPClass(WriteChunks::class);
+  }
+
+  /**
+   * Use streaming class, which defaults to `WriteChunks`
+   *
+   * @param  string|lang.XPClass $streaming
+   * @return self
+   */
+  public function using($streaming) {
+    $this->streaming= $streaming instanceof XPClass ? $streaming : XPClass::forName($streaming);
+    return $this;
+  }
 
   /**
    * Begins a request
@@ -19,6 +38,9 @@ class TestOutput extends Output {
     }
     $this->bytes.= "\r\n";
   }
+
+  /** @return web.io.Output */
+  public function streaming() { return $this->streaming->newInstance($this); }
 
   /**
    * Writes the bytes (in this case, to the internal buffer which can be